unimart-bot/requirements.txt
flask
flask-cors
Flask-Compress
python-dotenv
google-generativeai
faiss-cpu
numpy
scikit-learn
PyMuPDF
tqdm